Buffix

+++ src/controllers/videoController.js
@@ -8,7 +8,6 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id).populate('owner');
-  console.log('video: ', video);
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found' });
   }
@@ -16,9 +15,15 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
   const video = await Video.findById(id);
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   res.render('edit', { pageTitle: `Edit: ${video.title}`, video });
 };
@@ -26,8 +31,14 @@ export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
+  const {
+    user: { _id },
+  } = req.session;
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -58,12 +69,21 @@ export const postUpload = async (req, res) => {
     user.save();
     return res.redirect('/');
   } catch (error) {
-    console.log(error);
     return res.status(404).render('upload', { pageTitle: 'Upload Video', errorMessage: error._message });
   }
 };
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const {
+    user: { _id },
+  } = req.session;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.render('404', { pageTitle: 'Video not found' });
+  }
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect('/');
 };
@@ -77,6 +97,5 @@ export const search = async (req, res) => {
       },
     });
   }
-  console.log(videos);
   return res.render('search', { pageTitle: 'Search', videos });
 };
